+++ firmware/CHANGELOG.md
@@ -34,6 +34,7 @@ Release template (copy/paste this for new release):
 
 ### Fixed
  - overdwell protection causes Null-pointer deref and crash #9435
+ - huge print in Lua walks into memory #10159
 
 ## August 2026 "Day 1631"